+++ Assets/Scripts/Utils/Utils.cs
@@ -3,7 +3,7 @@
  */
 public static class Utils
 {
-    public static int Clamp(int value, int min, int max)
+    public static float Clamp(float value, float min, float max)
     {
         return (value < min) ? min : (value > max) ? max : value;
     }
